+++ src/graphcreation/collisionbased/message/ConnectServicesMessage.java
@@ -78,7 +78,7 @@ public class ConnectServicesMessage extends RemoteMessage implements PayloadMess
 
 	private void readMap(Map<Service, Set<ServiceDistance>> map, ObjectInput in) throws ClassNotFoundException, IOException {
 		List<Service> keys = Arrays.asList((Service[])in.readObject());
-		int size = in.readInt();
+		short size = in.readShort();
 		List<Set<ServiceDistance>> values = new ArrayList<Set<ServiceDistance>>();
 		for (int i = 0; i < size; i++) {
 			Set<ServiceDistance> value = new HashSet<ServiceDistance>(Arrays.asList((ServiceDistance[])in.readObject()));
@@ -98,7 +98,7 @@ public class ConnectServicesMessage extends RemoteMessage implements PayloadMess
 
 	private void writeMap(Map<Service, Set<ServiceDistance>> map, ObjectOutput out) throws IOException {
 		out.writeObject(map.keySet().toArray(new Service[0]));
-		out.writeInt(map.values().size());
+		out.writeShort(map.values().size());
 		for (Set<ServiceDistance> set : map.values())
 			out.writeObject(set.toArray(new ServiceDistance[0]));
 	}
